Moved ReadPattern into separate module
